src/app/requirements.txt
fastapi
python-multipart
uvicorn
ultralytics
pillow
numpy>=1.23.5
opencv-python>=4.1.1
PyYAML>=5.3.1
requests>=2.32.2
scipy>=1.4.1
torch>=1.8.0
torchvision>=0.9.0
tqdm>=4.66.3
pandas>=1.1.4
packaging